Update schema.md
Verify that the bolded "Paypal" instances aren't supposed to be "PayPal." Also, consider expanding on the definitions on some of the lower tables. For instance, in attributes labeled "merchantLocalDate", consider putting, "The merchant's local date in ISO 8601 format." This will help stay consistent with the other tables and attribute definitions.